Updated methods in facade to the way Thomas has done it.

+++ src/main/java/facades/SearchFacade_Impl.java
@@ -5,6 +5,7 @@ import dto.CityInfoDTO_OUT;
 import dto.HobbyDTO_IN;
 import dto.PersonDTO_IN;
 import dto.PersonDTO_OUT;
+import dto.PhoneDTO_IN;
 import entities.Address;
 import entities.CityInfo;
 import entities.Hobby;
@@ -127,13 +128,8 @@ public class SearchFacade_Impl implements ISearchFacade {
 
     @Override
     public PersonDTO_OUT editPerson(PersonDTO_IN personDTO) {
-        // Guard for PersonDTO being null / empty
-        if (personDTO == null || personDTO.getEmail() == null || personDTO.getEmail().isEmpty()
-                || personDTO.getFirstName() == null || personDTO.getFirstName().isEmpty()
-                || personDTO.getLastName() == null || personDTO.getLastName().isEmpty()
-                || personDTO.getAddress() == null || personDTO.getHobbies() == null || personDTO.getHobbies().isEmpty()
-                || personDTO.getPhones() == null || personDTO.getPhones().isEmpty()) {
-            throw new WebApplicationException("Missing Input", 400);
+        if (personDTO.getId() == 0) {
+            throw new WebApplicationException("Person ID was not set.", 400);
         }
 
         EntityManager em = getEntityManager();
@@ -141,29 +137,88 @@ public class SearchFacade_Impl implements ISearchFacade {
         try {
             em.getTransaction().begin();
             // Get person.
-            Person person = em.find(Person.class, personDTO.getId());
+            Person person_database = em.find(Person.class, personDTO.getId());
 
-            // Merge city.
-            em.merge(person.getAddress().getCityinfo());
+            if (person_database == null) {
+                throw new WebApplicationException("Could not find Person. Provided ID does not exist.", 400);
+            }
 
-            // Merge Address
-            em.merge(person.getAddress());
+            // Change all fields from the given DTO.
+            // Address
+            Address address;
+            if (personDTO.getAddress() != null) {
+                address = em.find(Address.class, personDTO.getAddress().getId());
+                if (address == null) {
+                    address = new Address(personDTO.getAddress());
+                    em.persist(address);
+                    person_database.setAddress(address);
+                } else {
+                    address = em.merge(address);
+                    person_database.setAddress(address);
+                }
+            }
 
-            // Merge Phones
-            person.getPhones().forEach((phone) -> {
-                em.merge(phone);
-            });
+            // First Name
+            if (personDTO.getFirstName() != null && !personDTO.getFirstName().isEmpty()) {
+                person_database.setFirstName(personDTO.getFirstName());
+            }
 
-            // Merge Hobbies
-            person.getHobbies().forEach((hobby) -> {
-                em.merge(hobby);
-            });
+            // Last Name
+            if (personDTO.getLastName() != null && !personDTO.getLastName().isEmpty()) {
+                person_database.setLastName(personDTO.getLastName());
+            }
 
-            // Merge Person
-            em.merge(person);
+            // Email
+            if (personDTO.getEmail() != null && !personDTO.getEmail().isEmpty()) {
+                person_database.setEmail(personDTO.getEmail());
+            }
 
-            return new PersonDTO_OUT(person);
+            // Hobbies
+            List<HobbyDTO_IN> hobbies = personDTO.getHobbies();
+            if (hobbies != null) {
+                hobbies.forEach((hobbyDTO) -> {
+                    Hobby hobby;
+                    hobby = em.find(Hobby.class, hobbyDTO.getId());
+                    if (hobby != null) {
+                        hobby.setDescription(hobbyDTO.getDescription());
+                        hobby.setName(hobbyDTO.getName());
+                        em.merge(hobby);
+                    } else {
+                        hobby = new Hobby(hobbyDTO);
+                        em.persist(hobby);
+                    }
+                    if (!person_database.getHobbies().contains(hobby)) {
+                        person_database.addHobby(hobby);
+                    }
+                });
+            }
 
+            // Phone
+            List<PhoneDTO_IN> phones = personDTO.getPhones();
+            if (phones != null) {
+                phones.forEach((phoneDTO) -> {
+                    Phone phone;
+                    phone = em.find(Phone.class, phoneDTO.getId());
+                    if (phone != null) {
+                        phone.setDescription(phoneDTO.getDescription());
+                        phone.setNumber(phoneDTO.getNumber());
+                        em.merge(phone);
+                    } else {
+                        phone = new Phone(phoneDTO);
+                        em.persist(phone);
+                    }
+                    if (!person_database.getPhones().contains(phone)) {
+                        person_database.addPhone(phone);
+                    }
+                });
+            }
+
+            Person mergedPerson = em.merge(person_database);
+            em.getTransaction().commit();
+            return new PersonDTO_OUT(mergedPerson);
+
+        } catch (RollbackException ex) {
+            throw new WebApplicationException("Something went wrong while editing person.", 500);
         } finally {
             em.close();
         }
@@ -176,6 +231,9 @@ public class SearchFacade_Impl implements ISearchFacade {
             em.getTransaction().begin();
             // Find Person from ID
             Person person = em.find(Person.class, id);
+            if (person == null) {
+                throw new WebApplicationException("Could not delete Person. Provided ID does not exist.", 400);
+            }
             // Remove all the phones
             person.getPhones().forEach((phone) -> {
                 em.remove(phone);
